Big Changes:
Changed the architecture from 32bits to 64bits. In order to be able to run the previous asm routine (which was in 32bits) in python (64bits) it was needed to convert and compile the asm code in 64 bits. Other changes: fixed the build.sh, small changes to the chart that result in a better view of the plot.